[move-model] separate translation out into its own file
